// app/Types/Character.ts
import LightCone from "./LightCone"

export default interface Character {
    additions: any
    attributes: any
    element:any
    icon: string
    id: string
    level: number
    light_cone: LightCone
    name: string
    path: any
    portrait: string
    preview: string
    promotion: number
    properties: any
    rank: number
    rarity: number
    relic_sets: any
    relics: any
    skills: any
}